Create README.md
The PATRONA puppy website is a website which is actually encyclopedia about all types of dogs. The Famous Veteran Patrona is fond of animals specially  Dogs, Cats, Rabbits. With her achievements in the field of serving as veteran, she also had interest in writing books on animals and their behavior.
it include user login and admin panel.